Relocate gui to new module Update setup.py entry points Cleanup resting screen implementation
