+++ neon_core/run_neon.py
@@ -121,7 +121,8 @@ def _stop_all_core_processes():
     procs = {p.pid: p.cmdline() for p in psutil.process_iter()}
     for pid, cmdline in procs.items():
         if cmdline and (any(pname in cmdline[-1] for pname in ("mycroft.messagebus.service", "neon_speech_client",
-                                                               "neon_audio_client", "neon_core.skills",
+                                                               "neon_audio_client",
+                                                               "neon_core.skills", "neon_core.gui"
                                                                "neon_core_server", "neon_enclosure_client",
                                                                "neon_core_client", "mycroft-gui-app",
                                                                "NGI.utilities.gui", "run_neon.py"))
@@ -162,6 +163,7 @@ def start_neon():
         _start_process("neon_enclosure_client")
         _start_process("neon_core_client")
         _start_process("mycroft-gui-app")
+        _start_process(["python3", "-m", "neon_core.gui"])
 
     try:
         STOP_MODULES.wait()
